+++ AimsProject/src/hust/soict/hedspi/test/cart/CartTest.java
@@ -1,6 +1,7 @@
-package hust.soict.hedspi.aims.cart;
+package hust.soict.hedspi.test.cart;
 
-import hust.soict.hedspi.aims.disc.*;
+import hust.soict.hedspi.aims.Cart.Cart;
+import hust.soict.hedspi.aims.disc.DigitalVideoDisc;
 
 public class CartTest {
     public static void main(String[] args) {
