+++ AimsProject/src/hust/soict/hedspi/test/store/StoreTest.java
@@ -1,6 +1,7 @@
-package hust.soict.hedspi.aims.disc;
+package hust.soict.hedspi.test.store;
 
-import hust.soict.hedspi.aims.store.*;
+import hust.soict.hedspi.aims.disc.DigitalVideoDisc;
+import hust.soict.hedspi.aims.store.Store;
 
 public class StoreTest {
     public static void main(String[] args) {
